requirements.txt
vllm>=0.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
Pillow>=10.0.0
huggingface_hub>=0.27.0
requests>=2.31.0
aiofiles>=24.1.0
